Restructure F3 docs and tests
