+++ features/F3/acceptance_tests/acceptance.py
@@ -1,0 +1,25 @@
+from .s1 import f3s1
+from .s2 import f3s2
+from .s3 import f3s3
+from .s4 import f3s4
+from .s5 import f3s5
+from .s6 import f3s6
+from .s7 import f3s7
+from .s8 import f3s8
+from .s9 import f3s9
+from .s10 import f3s10
+from .s11 import f3s11
+
+__all__ = [
+    "f3s1",
+    "f3s2",
+    "f3s3",
+    "f3s4",
+    "f3s5",
+    "f3s6",
+    "f3s7",
+    "f3s8",
+    "f3s9",
+    "f3s10",
+    "f3s11",
+]
